feat!: standardize functional chat room code
Standardize function chat room code to mimic and REST API
structure, implement authorization for new socket connections,
and restructure event messages.

Now the auth token and room Id is sent in socket header, so
in the connection the server will authorize the connection
and store the roomId and user information got from the token,
so the client do not have to send information
about roomId or user information into the messages.

BREAKING: breaks older code